+++ backend/statement-service/src/main/java/id/payu/statement/dto/TransactionType.java
@@ -1,4 +1,4 @@
-package id.payu.statement.application.service;
+package id.payu.statement.dto;
 
 public enum TransactionType {
         CREDIT, DEBIT